+++ src/pr_cmds.c
@@ -356,7 +356,7 @@ void PF_normalize (void)
 	value1 = G_VECTOR(OFS_PARM0);
 
 	new = value1[0] * value1[0] + value1[1] * value1[1] + value1[2]*value1[2];
-	new = sqrt(new);
+	new = sqrtf(new);
 	
 	if (new == 0)
 		newvalue[0] = newvalue[1] = newvalue[2] = 0;
@@ -386,7 +386,7 @@ void PF_vlen (void)
 	value1 = G_VECTOR(OFS_PARM0);
 
 	new = value1[0] * value1[0] + value1[1] * value1[1] + value1[2]*value1[2];
-	new = sqrt(new);
+	new = sqrtf (new);
 	
 	G_FLOAT(OFS_RETURN) = new;
 }
@@ -447,7 +447,7 @@ void PF_vectoangles (void)
 		if (yaw < 0)
 			yaw += 360;
 
-		forward = sqrt (value1[0]*value1[0] + value1[1]*value1[1]);
+		forward = sqrtf (value1[0]*value1[0] + value1[1]*value1[1]);
 		pitch = (int) (atan2(value1[2], forward) * 180 / M_PI);
 		if (pitch < 0)
 			pitch += 360;
@@ -1817,7 +1817,7 @@ void PF_cos (void)
 
 void PF_sqrt (void)
 {
-	G_FLOAT(OFS_RETURN) = sqrt(G_FLOAT(OFS_PARM0));
+	G_FLOAT(OFS_RETURN) = sqrtf (G_FLOAT(OFS_PARM0));
 }
 #endif
